requirements.txt
pandas
psycopg2
sqlalchemy
load_dotenv